requirements.txt
fastapi
uvicorn
pydantic
starlette
click
anyio
sniffio
websockets
python-multipart
twilio
python-dotenv
pydub
python-dateutil
six
httpx
pygments
rich
attrs
openai
PyJWT
tzdata
pinecone
langchain-openai
langchain-text-splitters
google-genai
